+++ packages/frontend/src/components/ProjectConnection/WarehouseForms/BigQueryForm.tsx
@@ -3,6 +3,7 @@ import {
     FeatureFlags,
     WarehouseTypes,
 } from '@lightdash/common';
+import type { SelectItem } from '@mantine/core';
 import {
     Anchor,
     Button,
@@ -21,6 +22,7 @@ import { IconCheck } from '@tabler/icons-react';
 import { useState, type ChangeEvent, type FC } from 'react';
 import { useToggle } from 'react-use';
 import { useGoogleLoginPopup } from '../../../hooks/gdrive/useGdrive';
+import useHealth from '../../../hooks/health/useHealth';
 import {
     useBigqueryDatasets,
     useIsBigQueryAuthenticated,
@@ -116,6 +118,8 @@ const BigQueryForm: FC<{
     const form = useFormContext();
     const project = form.getInputProps('warehouse.project');
     const [debouncedProject] = useDebouncedValue(project.value, 300);
+    const health = useHealth();
+    const isAdcEnabled = health.data?.auth.google?.enableGCloudADC;
 
     const isSsoEnabled = useFeatureFlagEnabled(FeatureFlags.BigquerySSO);
     // Fetching databases can only happen if user is authenticated
@@ -136,8 +140,11 @@ const BigQueryForm: FC<{
     const [isOpen, toggleOpen] = useToggle(false);
     const [temporaryFile, setTemporaryFile] = useState<File | null>(null);
     const { savedProject } = useProjectFormContext();
-    const requireSecrets: boolean =
-        savedProject?.warehouseConnection?.type !== WarehouseTypes.BIGQUERY;
+    const requireSecrets: boolean = !(
+        savedProject?.warehouseConnection?.type === WarehouseTypes.BIGQUERY &&
+        savedProject?.warehouseConnection?.authenticationType ===
+            BigqueryAuthenticationType.PRIVATE_KEY
+    );
     const hasDatasets = datasets && datasets.length > 0;
     const executionProjectField = form.getInputProps(
         'warehouse.executionProject',
@@ -170,10 +177,24 @@ const BigQueryForm: FC<{
     const isPassthroughLoginFeatureEnabled = useFeatureFlagEnabled(
         FeatureFlags.PassthroughLogin,
     );
+    const authenticationTypes = [
+        {
+            value: BigqueryAuthenticationType.PRIVATE_KEY,
+            label: 'Service Account (JSON key file)',
+        },
+        isSsoEnabled && {
+            value: BigqueryAuthenticationType.SSO,
+            label: 'User Account (Sign in with Google)',
+        },
+        isAdcEnabled && {
+            value: BigqueryAuthenticationType.ADC,
+            label: 'Application Default Credentials',
+        },
+    ].filter(Boolean) as SelectItem[];
     return (
         <>
             <Stack style={{ marginTop: '8px' }}>
-                {isSsoEnabled && (
+                {(isSsoEnabled || isAdcEnabled) && (
                     <Group spacing="sm">
                         <Select
                             name="warehouse.authenticationType"
@@ -199,16 +220,7 @@ const BigQueryForm: FC<{
                                     'Choose whether to authenticate with a service account or a user account'
                                 )
                             }
-                            data={[
-                                {
-                                    value: BigqueryAuthenticationType.PRIVATE_KEY,
-                                    label: 'Service Account (JSON key file)',
-                                },
-                                {
-                                    value: BigqueryAuthenticationType.SSO,
-                                    label: 'User Account (Sign in with Google)',
-                                },
-                            ]}
+                            data={authenticationTypes}
                             required
                             disabled={disabled}
                             w={isAuthenticated ? '90%' : '100%'}
@@ -332,7 +344,8 @@ const BigQueryForm: FC<{
                         }}
                 />         */}
                     </>
-                ) : (
+                ) : authenticationType ===
+                  BigqueryAuthenticationType.PRIVATE_KEY ? (
                     <>
                         <FileInput
                             name="warehouse.keyfileContents"
@@ -413,6 +426,9 @@ const BigQueryForm: FC<{
                             disabled={disabled}
                         />
                     </>
+                ) : (
+                    /* BigqueryAuthenticationType.ADC */
+                    <></>
                 )}
                 <FormSection isOpen={isOpen} name="advanced">
                     <Stack style={{ marginTop: '8px' }}>