+++ packages/frontend/src/components/UserCompletionModal/UserCompletionModal.tsx
@@ -99,7 +99,29 @@ const UserCompletionModal: FC = () => {
         setFieldValue('enableEmailDomainAccess', true);
     }, [canEnableEmailDomainAccess, setFieldValue]);
 
-    if (!user.data || user.data.isSetupComplete) {
+    useEffect(() => {
+        // Tracking is disabled, we complete the user tracking with default values
+        if (
+            health.data?.rudder.writeKey === undefined &&
+            !isLoading &&
+            user.data &&
+            !user.data.isSetupComplete
+        ) {
+            mutate({
+                organizationName: user.data.organizationName,
+                jobTitle: 'Other',
+                enableEmailDomainAccess: false,
+                isMarketingOptedIn: false,
+                isTrackingAnonymized: true,
+            });
+        }
+    }, [health.data?.rudder.writeKey, isLoading, mutate, user.data]);
+
+    if (
+        !user.data ||
+        user.data.isSetupComplete ||
+        health.data?.rudder.writeKey === undefined
+    ) {
         return null;
     }